requirements.txt
google-adk>=1.0.0
mcp>=1.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
